+++ functions.php
@@ -6,4 +6,6 @@ require_once get_template_directory() . '/inc/jetpack.php';
 require_once get_template_directory() . '/inc/search.php';
 require_once get_template_directory() . '/inc/posts.php';
 require_once get_template_directory() . '/inc/volumes.php';
+require_once get_template_directory() . '/inc/tags.php';
+require_once get_template_directory() . '/inc/more-link.php';