+++ patterns/template-single-post.php
@@ -2,7 +2,7 @@
 // Start the WordPress Loop
 while ( have_posts() ) : the_post();
 
-?>
+	?>
 
 	<article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
 		<header class="post-header">
@@ -14,41 +14,19 @@ while ( have_posts() ) : the_post();
 
 		<div class="post-content">
 			<?php
-				the_content();
+			the_content();
 
-				custom_post_pagination();
+			custom_post_pagination();
 
-				wp_link_pages( array(
-					'before'         => '<div class="post-pagination">' . __('<h5>Pagine</h5><p class="page-links">', 'textdomain'),
-					'next_or_number' => 'number',
-					'after'          => '</p></div>',
-				) );
+			wp_link_pages( array(
+				'before'         => '<div class="post-pagination">' . __('<h5>Pagine</h5><p class="page-links">', 'textdomain'),
+				'next_or_number' => 'number',
+				'after'          => '</p></div>',
+			) );
 			?>
 		</div>
 
-		<footer class="post-footer">
-			<?php if ( has_tag() ) : ?>
-			<div class="post-tags-list">
-				<h3>Argomenti Correlati</h3>
-				<p class="description">Esplora gli argomenti trattati in questo articolo: clicca su un’etichetta per leggere altri contenuti sullo stesso tema.</p>
-				<?php
-					the_tags(
-						'<ul class="post-tags"><li>',
-						'</li><li>',
-						'</li></ul>'
-					);
-				?>
-			</div>
-			<?php endif ?>
-			<div class="related-articles">
-			<?php
-				// Jetpack related posts (if available)
-				if ( class_exists( 'Jetpack_RelatedPosts' ) ) {
-					echo do_shortcode( '[jprel]' );
-				}
-			?>
-			</div>
-		</footer>
+		<?php get_template_part('patterns/template-single-post-footer'); ?>
 	</article>
 
 <?php endwhile; ?>